created useChartTooltip hook
