+++ src/components/Chart/LineWrapper/LineWrapper.module.scss
@@ -1,0 +1,5 @@
+@import 'src/styles/helpers';
+
+.chartWrapper {
+  position: relative;
+}
