+++ src/components/Chart/Tooltip/Tooltip.module.scss
@@ -1,9 +1,5 @@
 @import 'src/styles/helpers';
 
-.chartWrapper {
-  position: relative;
-}
-
 .tooltip {
   padding: 12px;
   position: absolute;
